requirements.txt
PyGitHub
PyYAML
dictdiffer
docker
gitpython
requests
ruamel.yaml
